requirements-tests.txt
docker-compose==1.14
flake8
mock
mypy
pytest
pytest-mock
tox
